rl_training/requirements.txt
numpy>=1.21.0
torch>=1.12.0
gym>=0.21.0
stable-baselines3>=1.6.0
jpype1>=1.4.0
tqdm>=4.64.0
